Updated snippet counter implementation
The idea is to have the backend as the keeper of the truth about
snippet counts. Therefore, each API request to the backend will bring
over the current personal and public snippet counters. The count will
be updated in the UI.

Also, the snippet counters will be taken our of the “topics” structure,
and given their own - “snippetCounters”